Revert "Sprint 05 (#2)"
This reverts commit e0afbfc200da6d2aeb154ffe5a234fcfae9fa24c.
