--- MovieQuiz/Models/AlertModel.swift
@@ -1,16 +1,0 @@
-//
-//  AlertModel.swift
-//  MovieQuiz
-//
-//  Created by Павел Врухин on 17.04.2023.
-//
-
-import Foundation
-
-struct AlertModel {
-    var title: String
-    var message: String
-    var buttonText: String
-    
-    var completion: () -> Void
-}
