--- MovieQuiz/Models/QuizQuestion.swift
@@ -1,14 +1,0 @@
-//
-//  QuizQuestion.swift
-//  MovieQuiz
-//
-//  Created by Павел Врухин on 14.04.2023.
-//
-
-import Foundation
-
-struct QuizQuestion {
-    let image: String
-    let text: String
-    let correctAnswer: Bool
-}
