--- MovieQuiz/Models/QuizResultViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  QuizResultsViewModel.swift
-//  MovieQuiz
-//
-//  Created by Павел Врухин on 14.04.2023.
-//
-
-import Foundation
-
-struct QuizResultViewModel {
-    let title: String
-    let text: String
-    let buttonText: String
-}
