--- MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  QuizStepViewModel.swift
-//  MovieQuiz
-//
-//  Created by Павел Врухин on 14.04.2023.
-//
-
-import UIKit
-
-struct QuizStepViewModel {
-    let image: UIImage
-    let question: String
-    let questionNumber: String
-}
