--- MovieQuiz/Presentation/AlertPresenterDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  AlertPresenterDelegate.swift
-//  MovieQuiz
-//
-//  Created by Павел Врухин on 18.04.2023.
-//
-
-import UIKit
-
-protocol AlertPresenterDelegate: AnyObject {
-    func didReceiveAlert(alert: UIAlertController?)
-}
