--- MovieQuiz/Services/QuestionFactoryDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  QuestionFactoryDelegate.swift
-//  MovieQuiz
-//
-//  Created by Павел Врухин on 16.04.2023.
-//
-
-import Foundation
-
-protocol QuestionFactoryDelegate: AnyObject {
-    func didReceiveNextQuestion(question: QuizQuestion?)
-}
